# resources/Wasteless_API_Spec.yml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: >-
      https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project team 800
  version: 6.0.0
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  /notifications:
    get:
      tags:
        - 'UCM4: Card expiry'
        - Cards / community marketplace
        - Notification
      summary: Retrieve the notifications of the logged in user
      parameters:
        - name: notificationType
          in: query
          description: The type of the notifications to find.
          schema:
            type: string
            example: CARD_DELETED
        - name: notificationTag
          in: query
          description: The tag of the notifications to find.
          schema:
            type: string
            example: MEDIUM_PRIORITY
        - name: notificationCategory
          in: query
          description: The category of the notifications to find.
          schema:
            type: string
            example: ARCHIVED
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: >-
            Request successful, a (possibly empty) list of notification is
            returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /notifications/{notificationId}:
    delete:
      tags:
        - 'UCM4: Card expiry'
        - Cards / community marketplace
        - Notification
      summary: (Custom) allows users to delete a notification
      parameters:
        - name: notificationId
          in: path
          description: The id of the notification to delete.
          required: true
          schema:
            type: integer
            example: 500
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: The notification was successfully deleted.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user who does not own a notification
            tries to delete it.
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  /notifications/{notificationId}/tag:
    put:
      tags:
        - 'U32: Managing my Feed'
        - Notification
      summary: (Custom) allows users to update a notifications tag
      parameters:
        - name: notificationId
          in: path
          description: The id of the notification to update
          required: true
          schema:
            type: integer
            example: 102
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationTagRequest'
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: The notification tag was successfully updated.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user who does not own a notification
            tries to update it.
  /notifications/{notificationId}/category:
    put:
      tags:
        - 'U32: Managing my Feed'
        - Notification
      summary: (Custom) allows users to update a notification's category
      parameters:
        - name: notificationId
          in: path
          description: The id of the notification to update
          required: true
          schema:
            type: integer
            example: 102
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCategoryRequest'
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: The notification category was successfully updated.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user who does not own a notification
            tries to update it.
...
  /notifications/message:
    post:
      tags:
        - 'UCM8: Contacting other marketplace users'
        - Cards / community marketplace
      summary: Inquire about another users card
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '201':
          description: Created the message
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID of the created message
                    example: 600
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
...
  /notifications/{notificationCacheId}/restore:
    summary: Restore a previously deleted notification by its id
    put:
      summary: Restore notification
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: notificationCacheId
          in: path
          description: id of notification to restore
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification has been restored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
      tags:
        - 'U32: Managing my feed'
        - Notification
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: >-
        There was some error with the data supplied by the user, appropriate
        error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
    ForbiddenError:
      description: The user does not have permission to perform the requested action
    NotAcceptableError:
      description: >-
        The requested route does exist (so not a 404) but some part of the
        request is not acceptable, for example trying to access a resource by an
        ID that does not exist.
...
  schemas:
    IdResponse:
      type: object
      description: Respresents an arbitrary id (e.g. of notification)
      properties:
        id:
          type: integer
          example: 42
          description: A unique id representing some object
...
    Notification:
      type: object
      description: Notification for event
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        recipientId:
          type: integer
          example: 100
          description: >-
            Recipient user id, if set to null then it will be system-wide
            notification
        senderId:
          type: integer
          example: 5
          description: ID of the sender
        message:
          type: string
          example: Hello World!
          description: Depends on the notification, each one has a different template
        type:
          type: string
          example: CARD_EXPIRING
          description: Depends on the notification, each one has a different template
        cardId:
          type: integer
          example: 1023
          description: ID of the card
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:32:00Z'
          readOnly: true
          description: Date the notification was created
        keywordId:
          type: integer
          example: 34
          description: KeywordId
    NotificationTagRequest:
      type: object
      description: Tag type for notification
      properties:
        notificationTagType:
          type: string
          example: 'High Priority'
          description: Tag type, high priority, low, medium, interesting, requires attention
    NotificationCategoryRequest:
      type: object
      description: Category type for notification
      properties:
        notificationCategory:
          type: string
          example: 'STARRED'
          description: UNREAD, READ, STARRED, ARCHIVED
    MessageRequest:
      type: object
      description: Message request for another user
      properties:
        cardId:
          type: integer
          example: 2
          description: ID of the card to inquire about
        recipientId:
          type: integer
          example: 2
          description: ID of the recipient (Either card owner, or replying to person)
        message:
          type: string
          example: Hello World!
          description: Associated message